AArch32: include platform_def.h in misc_helpers.S
When compiling for AArch32 i.MX7 we encounter "Error: selected processor
does not support `smc #0' in ARM mode" when compiling misc_helpers.S with
the Linaro GCC 7.2.1 cross-compiler.

This error can be fixed by declaring the architectural extension
".arch_extension sec".

The options for adding this extension are

- Directly in misc_helpers.S
- Inside of a platform-specific header

This patch chooses the latter solution so that platforms can more easily
manipulate this and other similar issues by way of the header.

A parallel patch for the NXP WaRP7 ATF port makes the follow declaration in
platform_def.h

/* Declare architectural extension sec to enable "smc #0" to compile */
.arch_extension sec

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>