imx7: Add stub warp7_bl2_el3_setup.c
On the i.MX7 the BootROM hands off to the second stage at EL3. The BootROM
is not ATF compliant though it does have a proprietary root-of-trust
authentication system that makes it possible to run ATF as an authenticated
binary.

This patch stubs out the necessary mandatory functions with some comments
for later binding to the lower-level initialization routines to come.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>