drivers: nxp: uart: Add stub mxc_console.S
This is enough to compile up an image but currently doesn't have any actual
code to waggle the UART in anyway.

This is the assembler version of this driver. The HAB hands off DRAM
already initialized so we may not in fact need to write our UART driver in
asm.

This is still to be determined.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>