drivers: nxp: uart: Add stub mxc_console
This is a stub console driver written in C. It is enough to compile up a
basic ATF image but doesn't actually try to write anything to the UART yet.

Various routines need to run before the console to set this up, clock setup
being the most obvious.

Given the HAB hands off to the next stage with the DRAM already initialized
it may be possible to just use the C version of the UART driver, this
driver, rather than bust out to assembly.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>