warp7: Fix "smc #0" compilation for warp7
A previous commit modified misc_helpers.S to include platform_def.h. That
modification allows us to define ".arch_extension sec" in a platform and
therefore processor specific way.

Declaring the ".arch_extension" in the header for the platform means the
person performing the porting activity needs to make the appropriate
declarations in their platform header, rather than say making an
".arch_extension sec" declaration in misc_helpers.S which could spuriously
declare an extension for a port that does not in-fact support it.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>