imx: aips: Add initial AIPS support
This patch adds an initial AIPS initialization routine. Setting up the
AIPSTZ controller is required to inform the SoC interconnect fabric which
bus-masters can read/write and if the read/writes are buffered.

For our purposes the initial configuration is for everything to be open. We
can lock-down later on as necessary.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>