imx: clock: Define clock identifiers
An array of clock gates is provided in the CCM block. In order to index
that array and thus enable/disable clock gates for the right components, we
need to provide meaningful names to the indices.

Section 5.2.5 of the i.MX7 Solo Application Processor Reference Manual Rev
0.1 provides the relevant details.

This patch documents those indices as an enum for later use.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>