imx: clock: Provide a clock gate enable/disable routine
This patch provides a clock-gate enable/disable routine via the set/clr
registers in a given clock-gate control register block.

This index passed should be one of the enums associated with CCM and
depending on enable/disable being passed either set or clr will be written
to.

The Domain0 bits are currently the only bits targeted by this write, more
work may need to be done on the domain bits in subsequent patches as a
result.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>