imx: clock: Define target mux select bits
This is a comprehensive definition of the target clock mux select bits.
These bits are required to correctly select the clock source. Defining all
of the bits up-front even for unused blocks in ATF means we can switch on
any block we want at a later date without having to write new code in the
clock-mux layer.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>