drivers: nxp: uart: Add code to perform initialization of UART
This patch initializes the MXC UART, take a crude method of calculating the
BAUD rate generator, not the UART clock-gates must have been enabled prior
to launching the UART init code.

Care needs to be taken to ensure the UBIR is initialized before the UBMR
and we need to ensure that UCR2.SRST comes good before trying to program
other registers associated with the UART - though at four cycles, the
likelihood of a racy-write is virtually zero.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>