drivers: nxp: uart: Introduce delay after CR2 reset
Currently we see invalid bytes output by the UART subsequent to resetting
via CR2. Introducing a delay after reset mitigates the invalid output.

Technically there should be no delay required, in practice it appears
CR2.SRST does not adequately gate the UART reset.

This patch introduces a delay after the SRST bit becomes true which appears
to fix the problem. An exact set of write sequences should be communicated
to NXP to elicit some feedback on why we see the behaviour identified and
ask for guidance on how to excise the delay.

For now keep the delay in-place and return to this issue at a later time.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>