warp7: io_storage: Add the struct io_policy structure
A standard means of describing how to load images is via the struct
io_policy structure. Oddly this structure has been copy/pasted a number of
times in various platform ports.

A big TOOD: here is to rationalize that down to one declaration and
propagate to the various users of it.

This patch replicates the definition and should go away in a rebase which
takes care of defining struct io_policy {}; correctly in some future patch.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>