warp7: io_storage: Add support for parsing a FIP via memory-map
This patch adds support to the io_storage layer of the WaRP7 port to
support parsing of a FIP in-memory.

This relies on a previous stage in the boot process to have located the FIP
at a known address.

That behaviour is very beneficial to us when debugging as we boot u-boot
first and then use u-boot to load ATF and FIP.

For production we will boot from eMMC but, its important to keep the
ability to-do u-boot -> {ATF, FIP} -> u-boot for ongoing development
purposes.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>