warp7: io_storage: Define io_block_storage for FIP
This commit defines the base and extent of FIP in memory for the case where
we have pre-loaded the FIP into memory and wish to memory map it.

During development we support BootROM-> u-boot -> ATF -> FIP

In this case u-boot will load ATF and FIP and will hand off control to ATF.
ATF will subsequently memmap() the FIP in memory and parse.

In the deployment version of ATF for WaRP7 this flow won't be used but, in
the initial stages it makes sense to leverage u-boot to load data out of
eMMC for ATF.

Later ATF patches will do the job of relocating data from eMMC to DRAM.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>