warp7: io_storage: Define open_memmap() for FIP in memory
Using u-boot to bootstrap ATF in early stage development means we want to
memmap() a FIP image already loaded into DRAM by u-boot.

The open_memmap() routine is pretty standard across platforms that
implement it and is itself deserving of a patch to rationalize the
implementation of this function throughout the various ATF ports.

This patch replicates open_memmap() with a view to superseding the
open_memmap() implementation with something to read the FIP out of eMMC
once we are producing eMMC images with the FIP embedded inside.

For now we'll let u-boot do the heavy lifting and map the FIP it has placed
in RAM for us.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>