warp7: io_storage: Define a FIP_IMAGE_ID entry for memmap() loading
In order to map the FIP already in memory when booting via u-boot we should
make an entry in a "struct plat_io_policy" variable.

This structure defines a callback to go and open the FIP.

The bl2_mem_params_desc builtin to the ATF image has a parameter that goes
looking for a BL2_IMAGE_ID.

This index gets passed to plat_get_image_source and the index is used to
lookup the policies[] array. This then triggers the opening of the FIP.

bl2_mem_params_descs ->
    plat_io_policy[] BL2_IMAGE_ID -> open_fip()
    plat_io_policy() FIP_IMAGE_ID -> memmap_fip()

In later patches we will replace the memory-mapped FIP open routine with an
eMMC/block-read routine by switching out the callback associated with
FIP_IMAGE_ID here.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>