warp7: io_storage: Provide an implementation of plat_get_image_source
plat_get_image_source() returns a handle to an image_spece and a handle to
a device on success or asserts on error. The implementation appears to be
completely standard across several platforms and is another candidate for
reductio-ab-absurdum sharing of code across platforms.

For now make another bog-standard implementation like the other
implementations and mark the necessary reduction work as TODO.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>