warp7: io_storage: Define open_fip()
This patch implements open_fip() for WaRP7. This appears to be a standard
function regularly re-implemented across ports and should be looked into as
a reductio-ab-absurdum activity.

open_fip() attempts to open an image from an indicated storage medium with
the FIP_IMAGE_ID passed as a parameter. This routine will return zero on
success or non-zero on failure.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>