warp7: io_storage: Define a BL32_IMAGE_ID policy entry
Define a BL32_IMAGE_ID entry in the io_storage policy array. When the
in-memory bl2_mem_params_desc calls into plat_get_image_source() for the
first time it will be attempting to load an image of type BL32_IMAGE_ID.

This will then trigger an attempted FIP open, which will in turn trigger a
routine to load the FIP.

In the first pass we are doing ATF development with u-boot loading ATF and
FIP, so the search of BL32_IMAGE_ID and subsequently FIP_IMAGE_ID will
trigger a memmap() of a pre-existing image and all going well the BL32 will
be located inside of that image.

Later on when we have SD images with ATF embedded inside of them we will
switch to reading FIP directly from the SD image in ATF either way this
section of code should remain as-is.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>