warp7: mem_params_desc: Add BL33 entry to mem params array
In order to both correctly load images out of the FIP and to pass
parameters to those images from ATF, we need a memory parameters entry in
bl2_mem_params_descs.

For the case of BL33 we should set the .next_handoff_image_id =
INVALID_IMAGE_ID to indicate we do not expect to try to handle any other
image types after BL33.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>