warp7: io_storage: Define a BL33_IMAGE_ID policy entry
Define a BL33_IMAGE_ID entry in the io_storage policy array. When the
in-memory bl2_mem_params_desc calls into plat_get_image_source() for the
first time it will be attempting to load an image of type BL32_IMAGE_ID, if
successful the next step is to load an image of type BL33_IMAGE_ID.

This patch adds the necessary BL33_IMAGE_ID policy descriptor to the
io_storage layer. In our case BL33_IMAGE_ID points to a u-boot image.

In a similar fashion to the BL32_IMAGE_ID addition, it is not of relevance
if we have used memmap() or eMMC to load the FIP into memory.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>