warp7: mem_params_desc: Add BL32_EXTRA2 entry to mem params array
In order to both correctly load images out of the FIP and to pass
parameters to those images from ATF, we need a memory parameters entry in
bl2_mem_params_descs.

This adds BL32_EXTRA2_IMAGE_ID to that array. BL32_EXTRA2 corresponds to
the pageable image, which in our case is a zero-sized binary.

The underlying OPTEE parsing code requires two valid pointers to be passed
into the validation routine BL32_EXTRA1 and BL32_EXTRA2.

This patch adds BL32_EXTRA2 to satisfy the requirement.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>