imx: io_mux: Define IO-mux bit-offsets consistent with docs
This patch defines the full range of IO-mux register offsets relative to
the base address of the IO-mux block base address.

We need the IO-mux stuff to setup the UART1 registers on the i.MX7 when
booting from the BootROM. Rather than make this a i.MX7 specific set of
defines, it looks as though these register offsets are shared with i.MX6
for example, so locate this file in plat/nxp/common instead of in an i.MX7
specific directory.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>