imx: io_mux: Define mux modes for UART1 TX/RX pads
This patch defines the bits for muxing the UART1 TX/RX lines. We could in
theory define all of the potentially mux bits for all of the registers but,
that would be a huge and costly venture.

For now just define the UART1 TX/RX bits and add more pad mux definition
bits in as needed.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>