imx: snvs: Define the bits of the HPCOMR register
A permission fix will need to be applied to the SNVS block prior to
switching on TrustZone. All we need to do is waggle a bit in the HPCOMR
register. To do that waggle we first need to define the bits of the HPCOMR
register.

This patch makes the necessary bit-definitions now.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>