drivers: nxp: crash-console: Implement mxc_crash_console_init
This patch implements the crash console UART init in assembly, as a
hard-coded 115200 8N1 assumed from the 24 MHz clock.

If the clock setup code has not run yet, this code can't work but, setting
up clocks and clock-gates is way out of scope for this type of recovery
function.

This patch, sets up the UART largely as the C version does, except it does
not support calculation of the BMR value as this requires performing a
division operation that requires stack, so instead of  a "bl
__aeabi_uidiv" we just hard-code the necessary value for the BMR @ 115200
clocked @ 24MHz.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>