drivers: nxp: crash-console: Rename asm version to mxc_crash_console
Looking at the code on the crash console path its clear we should be
providing a non-stack based UART init function for crash console.

The caveat will be that we will need to hard-code the crash console to
115200 8N1 or re-implement __aeabi_uidiv without a stack.

Doing that re-implementation while fun, is certainly out of scope, so,
we'll hard-code to 115200.

This patch converts the names in mxc_console.S to indicate they will be
used in a crash console.

Later on we can squash this patch down into the original patch creating the
file to begin with.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>