drivers: nxp: fsl_mmc: Change "error" to "state" in printout
Change the string "error" to "state" in the printout of errors. Later
patches will include a printout of an error variable indicating values such
as -EIO or -ETIMEDOUT.

This patch then changes up the string value so that the later patches can
have a better association of error-text to printed value.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>