imx7: imx-regs: bootrom: Define the absolute base address of the BootROM
The i.mx6, i.mx7, i.mx28 and upcoming i.mx9 processor families provide a
BootROM API callback for the purposes of High Assurance Boot (HAB).

The HAB function pointer table resides at a processor-specific offset from
the BootROM base address.

The first step in mapping that BootROM API is to define the base address
from which to offset from - in absolute physical address-space terms.

In this case the base address of the BootROM is @ 0x00000000.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Ryan Harkin <ryan.harkin@linaro.org>