imx7: hab_arch: Define the offset of the HAB API on i.MX7
This patch defines the offset from the BootROM base address of the HAB
function pointer table.

We can latch a set of common function pointer locations from the offset
given here and if necessary change the offset for different processors
without any other code-change.

For now all we support is i.MX7 so the only offset being defined is that
for the i.MX7 here.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Reviewed-by: Ryan Harkin <ryan.harkin@linaro.org>