warp7: Allow for selection between different MMC base addresses
To facilitate booting the same image on WaRP7 and SABRE7 we need to do a
little bit of boot-path frigging.

In this case we need to be able to direct to boot from the first SD device
as opposed to the third.

To facilitate the boot selection introduce a define that allows
redirection. If the define is not passed by the build just default to the
last SD device and continue.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>