warp7: system-counter: Program and enable system counter
This patch programs and enables the system counter @ 8MHz. This is
accomplished by first writing into the frequency mode table by way of an
mcr instruction.

Second a write to the optional CNTCR register enables the counter.

Using both the counter value and the counter frequency software can use the
tick to calculate delays.

Writing into CNTCR can only be done @ PL1, hence we need to initialize this
register prior to handing over to u-boot.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>