imx7: imx-regs: caam: Define base-address, job-ring, RTIC and DECO details
This patch defines the base address of the CAAM in the AIPS3 address range,
the CAAM_NUM_JOB_RINGS, the number of Real-Time-Integrity-Check (RTIC)
blocks and the number of descriptor-controllers (DECO) which are
responsible for executing CAAM descriptors.

Taken in aggregate then this patch specifies the SoC specific semantics of
the CAAM for the i.MX7 allowing later patches to define shared CAAM
semantics that ought to be useful for i.MX6, i.MX7, i.MX25 and i.MX8 should
any of those processors end up doing TF-A ports in the future.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>