imx7: caam: Define the base address of the CAAM in the AIPS3 range
This patch defines the base address of the CAAM in the AIPS3 address range.
The CAAM appears twice in the memory map, firstly inside the address space
of the A7 directly and confusingly then inside the AIPS/SoC address space.

The A7 address space is where the CAAM SRAM lives, whereas the AIPS address
range is where the CAAM registers live.

This patch makes the appropriate base address reference to the registers.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>