imx7: caam: Define the number of job rings available
The i.MX7 has three job rings available for CAAM job submission. The number
of job-rings can vary from processor to processor so, we should define how
many rings are available depending on the type of processor we are dealing
with.

This patch defines CAAM_NUM_JOB_RINGS as 0x03. Later this value will be
used both for defining the caam_ctrl structure and for looping through the
job-rings in that structure and assigning permission ownership per
job-ring.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>