imx7: caam: Define the number of RTIC blocks
The run-time-integrity-checker (RTIC) is a function of the CAAM block which
allows for simultaneous checking of up to four memory blocks.

The number of RTIC blocks seems designed to be variable in theory if not in
practice from processor-to-processor.

For our purposes we will just define the number of RTIC registers to map
and move on, no RTIC interaction is envisaged for the ATF port.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>