imx7: caam: Define the number of DECO blocks
The descriptor-controller (DECO) is responsible for executing CAAM
descriptors. Similarly to the job-ring and real-time-integrity-check
registers the DECO takes the form of a pair of 32-bit registers specifying
high-order and low-order bits. Again several instances of the DECO can in
theory be instantiated inside the CAAM block.

For the i.MX7 there is but one DECO set. Define the number of DECO blocks
so that the generic CAAM header can simply calculate rather than hard-code.

Slightly more effort, probably not worth the effort but, the right way to
define a variable structure within a over-all fixed address space.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>