imx: caam: Define the basic CAAM control structure
This patch defines the most basic part of the CAAM and the only piece of
the CAAM silicon we are really interested in, in ATF, the CAAM control
structure.

The CAAM itself is a huge address space of some 32k, way out of scope for
the purpose we have in ATF.

For now define the CAAM control structure only and should it become
necessary for some unforeseen requirement to implement a full-fat CAAM we
can extend and define this out further as necessary.

The current objective is to initialize the job-ring ownership to non-secure
world for later phases of the boot process to decide on appropriate
ownership.

Defining the CAAM control structure in full is actually over-kill for that
use-case. Define the structure now in anticipation of the upcoming job-ring
init.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>