warp7: caam: Add in the CAAM
This patch adds the CAAM into the build and calls into the previously
introduced caam_init() function.

The CAAM init done here in ATF is extremely limited, in that only the
job-ring ownership is assigned.

It may be necessary later to do a more comprehensive blitz of the CAAM but,
for now job-ring assignment appears all we need to care about.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>