imx7: wdog: Define the base addresses of the watchdog blocks
Booting up to Linux we get about 16 seconds into the boot and fall over.
Reviewing the i.MX7 Solo Application Processor Reference Manual we see in
section 6.5.1 the following text: "There is a power down counter which is
enabled out of any reset (POR, Warm/Cold). This counter has a fixed timeout
period of 16 seconds, upon which it asserts the WDOG signal.".

Clearly then it is necessary to map the watchdog registers and disable this
signal prior to handing off to the next phase of boot.

First step is mapping the base addresses within the AIPS1 block.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>