warp7: wdog: Perform minimal watchdog setup
This patch enables the watchdog clock-gates and switches off the watchdog
power-down-counter (PDE). The PDE is switched on, on-boot and if left in
the default boot-state will cause a processor reset after 16 seconds.

The accompanying u-boot operating from bl33 defines
CONFIG_SKIP_LOWLEVEL_INIT which means that u-boot never runs
imx_wdog_disable_powerdown() and so eventually we fall over on boot.

Bit twiddling the PDE from ATF is therefore required. This patch does the
necessary calls to make that happen.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>