bl2-el3: Fix exit to bl32 via eret
On two different sorts of i.MX7 boards the WaRP7 and PicoPi we have
observed occasional failure to hand over from BL2 to BL32 i.e. from ATF to
OPTEE.

Drilling down into the failure on the subset of boards we see the failure
subsequent to eret.

The root cause of this is that eret only populates the CPSR with the SPSR
when executing in Hyp Mode and not in Secure mode.

Populating CPSR from SPSR when NS = 0 with eret is not architecturally
defined.

This patch addresses the problem by specifically loading CPSR with the
required target mode, prior to executing eret.

Fixes: b1d27b484f41 ("bl2-el3: Add BL2_EL3 image")

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>